% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{Cross Validation for K-Nearest Neighbors}
\usage{
my_knn_cv(data, cl, k_nn, k_cv)
}
\arguments{
\item{data}{The input data frame used for prediction.
Note this data should not include NAs.}

\item{cl}{The true class values of the data, which are compared against the KNN prediction.
Note these classification should not include NAs.}

\item{k_nn}{An integer that represents the number of neighbors KNN should use.}

\item{k_cv}{The number of folds that the function will use for cross validation.}
}
\value{
A list containing the following items.
The vector \code{class}, which contains the predicted classes when KNN train on the entire provided data set.
The number \code{cv_err}, which is between 0 and 1 and shows the cross-validation misclassification error.
}
\description{
Run k-fold cross validations for k-nearest neighbors.
}
\keyword{inference}
