% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fitting Linear Models}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula class object that specifies the structure of the model.}

\item{data}{A data frame that contains the variables of the model.}
}
\value{
A matrix in the form of a table that contains an estimate, standard error, t-value, and p-value of each coefficient.
}
\description{
This function fits a linear model in the form of the specified formula using the given data.
}
\examples{
## Load in the penguins data set included in project3part1package
data(my_penguins)
my_penguins <- na.omit(my_penguins)

my_lm(bill_length_mm ~ bill_depth_mm + flipper_length_mm, data = my_penguins)
my_lm(body_mass_g ~ bill_length_mm - 1, data = my_penguins)

}
\keyword{inference}
\keyword{prediction}
