% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Cross Validation for Random Forests}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{The number of folds to use for cross validation.

@keywords inference

@return The average MSE across all k folds.}
}
\description{
This function runs k-fold cross validation for random forests using the penguins data set.
}
