% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{Cross Validation for K-Nearest Neighbors}
\usage{
my_knn_cv(data, cl, k_nn, k_cv)
}
\arguments{
\item{data}{The input data frame used for prediction.
Note this data should not include NAs.}

\item{cl}{The true class values of the data, which are compared against the KNN prediction.
Note these classification should not include NAs.}

\item{k_nn}{An integer that represents the number of neighbors KNN should use.}

\item{k_cv}{The number of folds that the function will use for cross validation.}
}
\value{
A list containing the following items.
The vector \code{class}, which contains the predicted classes when KNN train on the entire provided data set.
The number \code{cv_err}, which is between 0 and 1 and shows the cross-validation misclassification rate.
}
\description{
Run k-fold cross validations for k-nearest neighbors.
}
\examples{
## Load and set up the penguins data
data(my_penguins)
my_penguins <- na.omit(my_penguins)

penguins_no_NA <- na.omit(my_penguins)
my_data <- data.frame("bill_length" = penguins_no_NA$bill_length_mm,
    "bill_depth" = penguins_no_NA$bill_depth_mm,
    "flipper_length" = penguins_no_NA$flipper_length_mm,
    "body_mass" = penguins_no_NA$body_mass_g)
my_cl <- penguins_no_NA$species

my_knn_cv(data = my_data, cl = my_cl, k_nn = 2, k_cv = 5)
my_knn_cv(data = my_data, cl = my_cl, k_nn = 7, k_cv = 3)

}
\keyword{inference}
